--- client/next.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-  eslint: { ignoreDuringBuilds: true },
-  typescript: { ignoreBuildErrors: true },
-  webpack: (config) => {
-    config.resolve.symlinks = false;
-
-    return config;
-  },
-};
