+++ client/utils/apiClient.ts
@@ -1,4 +1,4 @@
 import aspida from '@aspida/axios';
 import api from 'api/$api';
 
-export const createApiClient = () => api(aspida(undefined, { baseURL: '/api' }));
+export const apiClient = api(aspida(undefined, { baseURL: '/api' }));
